convert PersistentCollection functional tests to unit tests
